CommentAPI working with Gateway, both on dev-auth branch. Authorization implemented on basic level but without roles yet

+++ Repositories/CommentRepository.cs
@@ -22,7 +22,7 @@ namespace BackEnd.Repositories
         {
             return await _context.Comments.ToListAsync();
         }
-        public async Task<IEnumerable<Comment>> GetAllCommentsInPostAsync(int postId)
+        public async Task<IEnumerable<Comment>> GetAllCommentsInPostAsync(long postId)
         {
             return await _context.Comments
                                  .Where(c => c.PostId == postId)
